mod6_labs/requirements.txt
flet==0.28.3
httpx
python-dotenv
OpenWeatherMap